+++ Source/Core/Common/ChunkFile.h
@@ -278,7 +278,7 @@ public:
     // bool's size can vary depending on platform, which can
     // cause breakages. This treats all bools as if they were
     // 8 bits in size.
-    u8 stable = static_cast<u8>(x);
+    u8 stable = x;
 
     Do(stable);
