+++ Source/Core/Core/HW/GCMemcard/GCMemcardRaw.cpp
@@ -64,7 +64,7 @@ MemoryCard::MemoryCard(const std::string& filename, const ExpansionInterface::Sl
     const CardFlashId& flash_id = sram.settings_ex.flash_id[Memcard::SLOT_A];
     const bool shift_jis = m_filename.find(".JAP.raw") != std::string::npos;
     const u32 rtc_bias = sram.settings.rtc_bias;
-    const u32 sram_language = static_cast<u32>(sram.settings.language);
+    const u32 sram_language = sram.settings.language;
     const u64 format_time =
         Common::Timer::GetLocalTimeSinceJan1970() - ExpansionInterface::CEXIIPL::GC_EPOCH;
     Memcard::GCMemcard::Format(&m_memcard_data[0], flash_id, size_mbits, shift_jis, rtc_bias,
